requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
bcrypt>=4.0.1
PyJWT>=2.8.0
email-validator>=2.1.0

# AI/ML
openai>=1.3.0
sentence-transformers>=2.2.2
tiktoken>=0.5.0

# Database
asyncpg>=0.29.0

# HTTP Client
aiohttp>=3.9.0
requests>=2.31.0

# Vector Store (optional - disabled for Railway by default)
# chromadb>=0.4.18

# File Processing
aiofiles>=23.2.0
PyMuPDF>=1.23.0
pillow>=10.0.0
numpy>=1.24.3
pdfplumber>=0.10.0
pandas>=2.0.0
pytesseract>=0.3.10

# Excel Processing
openpyxl>=3.1.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.20.0

# Text Processing
langchain-text-splitters>=0.0.1

# Testing
hypothesis>=6.92.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
